增加 PictureController

+++ src/main/java/com/penny/vo/base/PropertyPictureBaseVo.java
@@ -16,5 +16,5 @@ public class PropertyPictureBaseVo {
 
     private Long pictureId;
 
-    private Long propertyPictureOrder;
+    private Integer propertyPictureOrder;
 }
